requirements.txt
Flask==3.0.0
rembg==2.0.50
Pillow==10.1.0
Werkzeug==3.0.1
gunicorn==21.2.0
numpy==1.24.3